% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_metabolism.R
\name{plot.metabolism}
\alias{plot.metabolism}
\title{A Cat Function}
\usage{
\method{plot}{metabolism}(
  obj,
  pathway,
  dimention.reduction.type = "umap",
  dimention.reduction.run = T,
  size = 1
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
plot.metabolism()
}
\keyword{cats}
