% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metabolism.R
\name{sc.metabolism}
\alias{sc.metabolism}
\title{A Cat Function}
\usage{
sc.metabolism(
  countexp,
  method = "VISION",
  imputation = F,
  ncores = 2,
  metabolism.type = "KEGG"
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
sc.metabolism()
}
\keyword{cats}
