% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metabolism_Seurat.R
\name{sc.metabolism.Seurat}
\alias{sc.metabolism.Seurat}
\title{scMetabolism}
\usage{
sc.metabolism.Seurat(
  obj,
  method = "VISION",
  imputation = F,
  ncores = 2,
  metabolism.type = "KEGG"
)
}
\description{
scMetabolism
}
\examples{
sc.metabolism.Seurat()
}
\keyword{scMetabolism}
