% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metabolism.R
\name{sc.metabolism}
\alias{sc.metabolism}
\title{scMetabolism}
\usage{
sc.metabolism(
  countexp,
  method = "VISION",
  imputation = F,
  ncores = 2,
  metabolism.type = "KEGG"
)
}
\description{
scMetabolism
}
\examples{
sc.metabolism()
}
\keyword{scMetabolism}
